+++ src/app/modules/project/edit/edit.component.ts
@@ -64,6 +64,11 @@ export class EditComponent implements OnInit {
    */
   public invalidId: string;
 
+  /**
+   * Property to indicate whether the project is loading.
+   */
+  public projectLoading = true;
+
   constructor(
     private router: Router,
     private formBuilder: FormBuilder,
@@ -95,12 +100,16 @@ export class EditComponent implements OnInit {
       return;
     }
 
-    this.projectService.get(id).subscribe(
-      (result) => {
-        this.project = result;
-        this.collaborators = this.project.collaborators;
-      }
-    );
+    this.projectService.get(id)
+      .pipe(
+        finalize(() => this.projectLoading = false)
+      )
+      .subscribe(
+        (result) => {
+          this.project = result;
+          this.collaborators = this.project.collaborators;
+        }
+      );
   }
 
   public onClickSubmit(): void {
@@ -123,8 +132,12 @@ export class EditComponent implements OnInit {
 
     this.projectService
       .put(this.project.id, edittedProject)
-      .pipe(finalize(() => (this.submitEnabled = false)))
-      .subscribe((result) => {
+      .pipe(
+        finalize(() => {
+          this.submitEnabled = false;
+        })
+      )
+      .subscribe(() => {
         const alertConfig: AlertConfig = {
           type: AlertType.success,
           mainMessage: 'Project was succesfully updated',
@@ -134,6 +147,17 @@ export class EditComponent implements OnInit {
         this.alertService.pushAlert(alertConfig);
         this.router.navigate([`/project/overview`]);
       });
+  }
+
+  /**
+   * Method  which triggers when the cancel button is pressed.
+   * Redirects the user back to the project or the overview.
+   */
+  public onClickCancel(): void {
+    if (this.project == null) {
+      this.router.navigate(['project/overview']);
+    }
+    this.router.navigate([`project/details/${this.project.id}`]);
   }
 
   /**